+++ tests/debounce.rs
@@ -333,6 +333,29 @@ fn create_rename_overwrite_file() {
     }
 }
 
+// https://github.com/passcod/notify/issues/99
+#[test]
+fn create_rename_write_create() { // fsevents
+    let tdir = TempDir::new("temp_dir").expect("failed to create temporary directory");
+
+    sleep_macos(10);
+
+    let (tx, rx) = mpsc::channel();
+    let mut watcher: RecommendedWatcher = Watcher::new(tx, Duration::from_secs(DELAY_S)).expect("failed to create debounced watcher");
+    watcher.watch(tdir.mkpath("."), RecursiveMode::Recursive).expect("failed to watch directory");
+
+    tdir.create("file1");
+    tdir.rename("file1", "file2");
+    sleep(10);
+    tdir.write("file2");
+    tdir.create("file3");
+
+    assert_eq!(recv_events_debounced(&rx), vec![
+        DebouncedEvent::Create(tdir.mkpath("file2")),
+        DebouncedEvent::Create(tdir.mkpath("file3")),
+    ]);
+}
+
 #[test]
 fn write_rename_file() {
     let tdir = TempDir::new("temp_dir").expect("failed to create temporary directory");